SoftUniBootcampExam/requirements.txt
asgiref==3.4.1
Django==3.2.6
django-rest-multiple-models==2.1.3
djangorestframework==3.12.4
jsonfield==3.1.0
psycopg2-binary==2.9.1
pytz==2021.1
sqlparse==0.4.1
